ttps/Active-Directory/3_PrivEsc/GroupPolicyBackdoor/requirements.txt
ldap3-bleeding-edge >= 2.10.1.1337
smbprotocol[kerberos]
typer
pydantic
lxml
chardet
